+++ tests/base/test_cli.py
@@ -298,7 +298,7 @@ def test_transient_mode(test_release_version: str) -> None:
         "subtype = residential",
     ],
     f"files/{TEST_RELEASE_VERSION}/theme=buildings/type=building/"
-    f"{_generate_geometry_hash(geometry_box())}_b22759b51edd150209b17a03319daa6796f574478c5f4f79ae2efa62e65e3ba1.parquet",
+    f"{_generate_geometry_hash(geometry_box())}_b22759b5.parquet",
 )  # type: ignore
 @P.case(
     "Geometry WKT filter",
@@ -311,7 +311,7 @@ def test_transient_mode(test_release_version: str) -> None:
         "building",
     ],
     f"files/{TEST_RELEASE_VERSION}/"
-    "theme=buildings/type=building/09c3fc0471538594b784be7c52782837c7a26753c2b26097b780581fa0a6bfc6_nofilter.parquet",
+    "theme=buildings/type=building/09c3fc04_nofilter.parquet",
 )  # type: ignore
 @P.case(
     "Geometry GeoJSON filter",
@@ -324,7 +324,7 @@ def test_transient_mode(test_release_version: str) -> None:
         "building",
     ],
     f"files/{TEST_RELEASE_VERSION}/"
-    "theme=buildings/type=building/82c0fdfa2d5654818a03540644834d70c353e3f82f9d8f201c37420aeb35118e_nofilter.parquet",
+    "theme=buildings/type=building/82c0fdfa_nofilter.parquet",
 )  # type: ignore
 @P.case(
     "Geometry file filter",
@@ -337,7 +337,7 @@ def test_transient_mode(test_release_version: str) -> None:
         "building",
     ],
     f"files/{TEST_RELEASE_VERSION}/"
-    "theme=buildings/type=building/6a869bcfa1a49ade8b76569e48e4142bce29098815bf37e57155a18204f2bbbc_nofilter.parquet",
+    "theme=buildings/type=building/6a869bcf_nofilter.parquet",
 )  # type: ignore
 @P.case(
     "Geometry geocode filter",
@@ -350,7 +350,7 @@ def test_transient_mode(test_release_version: str) -> None:
         "building",
     ],
     f"files/{TEST_RELEASE_VERSION}/"
-    "theme=buildings/type=building/e7f0b78a0fdc16c4db31c9767fa4e639eadaa8e83a9b90e07b521f4925cdf4b3_nofilter.parquet",
+    "theme=buildings/type=building/e7f0b78a_nofilter.parquet",
 )  # type: ignore
 @P.case(
     "Geometry Geohash filter",
@@ -363,7 +363,7 @@ def test_transient_mode(test_release_version: str) -> None:
         "building",
     ],
     f"files/{TEST_RELEASE_VERSION}/"
-    "theme=buildings/type=building/c08889e81575260e7ea2bc9764ddaa7c5e1141270a890b022799689d39dfe4d5_nofilter.parquet",
+    "theme=buildings/type=building/c08889e8_nofilter.parquet",
 )  # type: ignore
 @P.case(
     "Geometry Geohash filter multiple",
@@ -376,7 +376,7 @@ def test_transient_mode(test_release_version: str) -> None:
         "building",
     ],
     f"files/{TEST_RELEASE_VERSION}/"
-    "theme=buildings/type=building/1bd33e0afc3cd0efcb4740185b8a05ecaf1bac916d571403768939b82844b43e_nofilter.parquet",
+    "theme=buildings/type=building/1bd33e0a_nofilter.parquet",
 )  # type: ignore
 @P.case(
     "Geometry H3 filter",
@@ -389,7 +389,7 @@ def test_transient_mode(test_release_version: str) -> None:
         "building",
     ],
     f"files/{TEST_RELEASE_VERSION}/"
-    "theme=buildings/type=building/a2f8d5114760394646aa999a1204adaa48ad686b3fcadb0b25fd02322c16dff4_nofilter.parquet",
+    "theme=buildings/type=building/a2f8d511_nofilter.parquet",
 )  # type: ignore
 @P.case(
     "Geometry H3 filter multiple",
@@ -402,7 +402,7 @@ def test_transient_mode(test_release_version: str) -> None:
         "building",
     ],
     f"files/{TEST_RELEASE_VERSION}/"
-    "theme=buildings/type=building/e50e6489d4faba664a3c7f9729b7a3bedafb7a396ae826521f32b556d0b554f1_nofilter.parquet",
+    "theme=buildings/type=building/e50e6489_nofilter.parquet",
 )  # type: ignore
 @P.case(
     "Geometry S2 filter",
@@ -415,7 +415,7 @@ def test_transient_mode(test_release_version: str) -> None:
         "building",
     ],
     f"files/{TEST_RELEASE_VERSION}/"
-    "theme=buildings/type=building/5c3d61eb108819e543a1a59fe6c67658f817c0453d728b5aa007f227453d5bf6_nofilter.parquet",
+    "theme=buildings/type=building/5c3d61eb_nofilter.parquet",
 )  # type: ignore
 @P.case(
     "Geometry S2 filter multiple",
@@ -428,7 +428,7 @@ def test_transient_mode(test_release_version: str) -> None:
         "building",
     ],
     f"files/{TEST_RELEASE_VERSION}/"
-    "theme=buildings/type=building/cda5d65e169e3ff04970e66e80b021937a5ae141ed72428cc0d9a3764bf076db_nofilter.parquet",
+    "theme=buildings/type=building/cda5d65e_nofilter.parquet",
 )  # type: ignore
 def test_proper_args(args: list[str], expected_result: str) -> None:
     """Test if runs properly with options."""